% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_nhanesR.R
\name{install_nhanesR}
\alias{install_nhanesR}
\title{install}
\usage{
install_nhanesR(token = "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")
}
\arguments{
\item{token}{token}
}
\value{
install
}
\description{
install
}
